requirements.txt
fastapi==0.113.0
uvicorn==0.30.6
pydantic==2.9.0
mysql-connector-python==9.0.0
